Dnialegen/requirements.txt
flask
gunicorn
request
random